[dds/dap] Don't show stack traces in error messages from DAP
We accidentally started showing the stack traces in all errors rendered in VS Code. This changes the format of the message to not include the stack - although it keeps it in the payload to aid debugging when capturing logs.

It also slightly tweaks the error message for global eval to be less cryptic for VS Code.

Fixes https://github.com/Dart-Code/Dart-Code/issues/4899

Change-Id: Id4a0c4e6ef26065059e3becec62bab5d873e7be3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/342621
Reviewed-by: Ben Konyi <bkonyi@google.com>
Reviewed-by: Helin Shiah <helinx@google.com>