[analyzer] Combine adding files + packages in declarations tracker to avoid enumerating analyzedFiles twice
Prior to this change, both `_scheduleContextFiles` and `_findPackages()`  would enumerate `_analysisContext.contextRoot.analyzedFiles()`. This is an Iterable that calls `getResource()` for every item so on large folders showed up on the profile:

https://github.com/dart-lang/sdk/issues/54513#issuecomment-1877065904

This change combines them into a single method so we iterate only once.

It does slightly change the order of things (SDK files were previously scheduled before packages), but the only alternative I have is to add `toList()` but that will need to allocate (and repeatedly grow) a list which I suspect is overall worse.

Other than the order, there should be no functional changes here - I moved the body of `_scheduleContextFiles` into `_findPackages()` and renamed it (which unfortunately doesn't diff nicely because it moved in the file).

See https://github.com/dart-lang/sdk/issues/54513

Change-Id: I2a5539a93ee8ca2cdbad031efcfc9bbaf23e9ab9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/344703
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>