[cfe] Show when syntax was unhandled in the error message
This updates the error reporting on incorrectly applied macro
application to show if an unimplemented for syntax was the cause.

This is to help identify when macro tests fail due to missing
implementation.

The CL includes handling of simple list literals in macro
annotations, an issue brought up this kind of error.

Change-Id: I4b6f9f7140c7bc000cfec54acbc6bbfeaf3429a2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/345662
Reviewed-by: Morgan :) <davidmorgan@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>