[beta][dart2js] Fix as-check type registration.
We would expect failing as checks to throw in all modes (unless --omit-as-casts is provided). However, the new test program fails in production mode. This is because we are not registering the type usage of the function's type parameter. This leads us to drop the as check completely later on.

We should be doing subtype checks with nullability if we want to consider an as test omitted.

Bug: https://github.com/dart-lang/sdk/issues/54514
Bug: https://github.com/dart-lang/sdk/issues/54419
Change-Id: I01269cb8f306799b518de8869f8264adc8594fae
Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/343821
Cherry-pick-request: https://github.com/dart-lang/sdk/issues/54556
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/345480
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nate Biggs <natebiggs@google.com>