Clean up mini-ast handling of `lub` and `promoteToNonNull`.
`lub` was implemented twice (once as `lub` and once as `_lub`), with
some behaviors implemented in each, and other behaviors not
implemented. This was good enough to make unit tests pass, but for
long term maintainability, it's better to have a single unified
implementation.

The unified implementation makes use of `promoteToNonNull`, which had
a subtle bug: it interpreted the unknown type `?` as a nullable type
and tried to convert it to ``. That bug has now been fixed.

Change-Id: If257a446e46f916843fe0ff95b6bc48c6c820d93
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/346421
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>