Record the completion location in more places
This is the first of many such CLs. We need to record the completion
everywhere that `OpType` is computing it in order to eventually be able
to stop using `OpType`. I have temporary code that compares the two
computations in order to ensure that they're the same, but I can't
commit it into prod, nor will it be useful beyond this transition.

In addition to recording the completion location in more places, this
also better handles the case when the offset is immediately before the
first character of a top-level declaration. I believe that we need to
do something similar when immedately after the last character, when
before and after class members, and when before an after statements.

Change-Id: I046675c56243981762cd06db4cd35c17989ccf27
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/348762
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>